Possible bug cleanup and general cleanup

+++ src/main/java/com/darkblade12/ParticleEffect/ParticleEffect.java
@@ -1147,7 +1147,7 @@ public enum ParticleEffect {
 		 */
 		@Override
 		public float getValueX() {
-			return (float) red / 255F;
+			return red / 255F;
 		}
 
 		/**
@@ -1157,7 +1157,7 @@ public enum ParticleEffect {
 		 */
 		@Override
 		public float getValueY() {
-			return (float) green / 255F;
+			return green / 255F;
 		}
 
 		/**
@@ -1167,7 +1167,7 @@ public enum ParticleEffect {
 		 */
 		@Override
 		public float getValueZ() {
-			return (float) blue / 255F;
+			return blue / 255F;
 		}
 	}
 
@@ -1205,7 +1205,7 @@ public enum ParticleEffect {
 		 */
 		@Override
 		public float getValueX() {
-			return (float) note / 24F;
+			return note / 24F;
 		}
 
 		/**
